requirements.txt
aioredis==1.3.1
amqp==5.0.6
argon2-cffi==20.1.0
asgiref==3.4.1
async-generator==1.10
async-timeout==4.0.2
attrs==21.2.0
autobahn==21.2.1
Automat==20.2.0
backcall==0.2.0
bcrypt==3.2.0
beautifulsoup4==4.11.1
billiard==3.6.4.0
bleach==4.0.0
blis==0.7.7
boto3==1.20.11
botocore==1.23.54
cachetools==4.2.2
catalogue==2.0.7
celery==5.1.2
certifi==2020.12.5
cffi==1.14.5
channels==3.0.3
channels-redis==3.3.1
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.4
constantly==15.1.0
cryptocode==0.1
cryptography==3.4.7
cymem==2.0.6
daphne==3.0.2
debugpy==1.4.1
decorator==5.0.7
defusedxml==0.7.1
Deprecated==1.2.13
dj-stripe==2.7.0a8
Django==3.2.12
django-allauth==0.44.0
django-celery-results==2.2.0
django-cors-headers==3.7.0
django-crispy-forms==1.11.2
django-dbbackup==3.3.0
django-extensions==3.1.3
django-filter==2.4.0
django-guardian==2.4.0
django-model-utils==4.2.0
django-notifications-hq==1.6.0
django-storages==1.12.3
djangorestframework==3.12.4
djangorestframework-simplejwt==5.1.0
docx2txt==0.8
drf-access-policy==1.1.1
drf-writable-nested==0.6.3
entrypoints==0.3
et-xmlfile==1.1.0
google-api-core
google-api-core[grpc,grpcgcp]
google-auth==1.34.0
google-cloud-core==1.7.2
google-cloud-speech==2.14.0
google-cloud-texttospeech==2.11.0
google-cloud-translate==3.2.0
googleapis-common-protos==1.53.0
graphviz==0.20
grpcio==1.46.1
grpcio-status==1.46.1
gunicorn==20.1.0
hiredis==2.0.0
hyperlink==21.0.0
idna==2.10
incremental==21.3.0
ipykernel==5.5.6
ipython==7.16.3
ipython-genutils==0.2.0
ipywidgets==7.6.3
jedi==0.17.0
Jinja2==3.0.1
jmespath==0.10.0
joblib==1.0.1
jsonfield==3.1.0
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.12
jupyter-console==6.4.0
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
kombu==5.1.0
langcodes==3.3.0
lxml==4.6.3
MarkupSafe==2.0.1
matplotlib-inline==0.1.2
mistune==0.8.4
msgpack==1.0.3
murmurhash==1.0.7
nbclient==0.5.3
nbconvert==6.0.7
nbformat==5.1.3
nest-asyncio==1.5.1
nltk==3.5
notebook==6.4.2
numpy==1.21
oauthlib==3.1.1
openpyxl==3.0.7
packaging==21.0
pandocfilters==1.4.3
parso==0.8.2
pathy==0.6.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.2.0
preshed==3.0.6
prometheus-client==0.11.0
prompt-toolkit==3.0.18
proto-plus==1.19.0
protobuf==3.17.3
psycopg2-binary==2.8.6
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pycryptodomex==3.14.1
pydantic==1.8.2
pydot==1.4.2
PyGithub==1.55
Pygments==2.9.0
PyJWT==2.1.0
pymongo==4.0.1
PyNaCl==1.5.0
pyOpenSSL==20.0.1
pyparsing==2.4.7
pyrsistent==0.18.0
pyspellchecker==0.6.2
python-dateutil==2.8.1
python-dotenv==0.17.1
python-gitlab==3.1.1
python3-openid==3.2.0
pytz==2021.1
pyzmq==22.2.1
qtconsole==5.1.1
QtPy==1.9.0
redis==4.3.1
regex==2021.4.4
requests==2.25.1
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
rsa==4.7.2
s3transfer==0.5.2
Send2Trash==1.8.0
sentry-sdk==1.5.12
service-identity==21.1.0
six==1.16.0
smart-open==5.2.1
soupsieve==2.3.2.post1
spacy==3.2.3
spacy-legacy==3.0.9
spacy-loggers==1.0.2
sqlparse==0.4.1
srsly==2.4.3
stripe==2.60.0
swapper==1.3.0
tablib==3.0.0
terminado==0.10.1
testpath==0.5.0
thinc==8.0.15
tornado==6.1
tqdm==4.62.0
traitlets==4.3.3
Twisted==21.2.0
txaio==21.2.1
typer==0.4.1
typing-extensions==3.10.0.0
urllib3==1.26.6
vine==5.0.0
wasabi==0.9.1
wcwidth==0.2.5
webencodings==0.5.1
widgetsnbextension==3.5.1
wiktionaryparser==0.0.97
wrapt==1.14.1
XlsxWriter==3.0.1
xlwt==1.3.0
graphviz
pydot
spacy==3.2.3
google-cloud-texttospeech
google-cloud-speech
gevent
zope.interface==5.4.0
docx2txt==0.8
weasyprint==55.0
gevent
indic-nlp-library==0.81
google.cloud.storage==1.44.0
mutagen
hunspell
htmldocx
quill-delta[html]
python-docx
langdetect==1.0.9
pydub
filesplit==4.0.1
